Disable IPNS and fixes for ipfs restart

+++ terraform/ipfs-server/process-stream.sh
@@ -41,49 +41,63 @@ while true; do
 
     # Grab the timecode from the m3u8 file so we can add it to the log
     timecode=`grep -B1 ${nextfile} ${what}.m3u8 | head -n1 | awk -F : '{print $2}' | tr -d ,`
-    attempts=10
-    until [[ "$timecode" || $attempts -eq 0 ]]; do
-      sleep 0.2
+    attempts=5
+    until [[ "${timecode}" || ${attempts} -eq 0 ]]; do
+      # Wait and retry
+      sleep 0.5
       timecode=`grep -B1 ${nextfile} ${what}.m3u8 | head -n1 | awk -F : '{print $2}' | tr -d ,`
       attempts=$((attempts-1))
     done
+    if ! [[ "${timecode}" ]]; then
+      # Set approximate timecode
+      timecode="${HLS_TIME}.000000"
+    fi
     
     # Calculate and store the offset needed if ffmpeg where to restart right now
     currentOffset=$(ffprobe ${nextfile} 2>&1 | grep Duration | awk '{print $4}' | tr -d ',')
-    nextOffset=$(echo "${currentOffset} + ${timecode}" |  bc)
+    nextOffset=$(echo "${currentOffset} + ${timecode}" | bc)
     echo ${nextOffset} > ~/stream-offset
 
     # Current UTC date for the log
     time=`date "+%F-%H-%M-%S"`
 
     # Add ts file to IPFS
-    ipfs add ${nextfile} > ~/tmp.txt
-    
-    # Update the log with the future name (hash already there)
-    echo $(cat ~/tmp.txt) ${time}.ts ${timecode} >> ~/process-stream.log    
-    
-    # Remove nextfile and tmp.txt
-    rm -f ${nextfile} ~/tmp.txt
+    ret=`ipfs add ${nextfile} 2>/dev/null > ~/tmp.txt; echo $?`
+    attempts=5
+    until [[ ${ret} -eq 0 || ${attempts} -eq 0 ]]; do
+      # Wait and retry
+      sleep 1
+      ret=`ipfs add ${nextfile} 2>/dev/null > ~/tmp.txt; echo $?`
+      attempts=$((attempts-1))
+    done
+    if [[ ${ret} -eq 0 ]]; then
+      # Update the log with the future name (hash already there)
+      echo $(cat ~/tmp.txt) ${time}.ts ${timecode} >> ~/process-stream.log
 
-    # Write the m3u8 file with the new IPFS hashes from the log
-    totalLines="$(wc -l ~/process-stream.log | awk '{print $1}')"
-    
-    sequence=0
-    if (( "${totalLines}" > ${M3U8_SIZE} )); then
-        sequence=`expr ${totalLines} - ${M3U8_SIZE}`
+      # Remove nextfile and tmp.txt
+      rm -f ${nextfile} ~/tmp.txt
+
+      # Write the m3u8 file with the new IPFS hashes from the log
+      totalLines="$(wc -l ~/process-stream.log | awk '{print $1}')"
+      
+      sequence=0
+      if (( "${totalLines}" > ${M3U8_SIZE} )); then
+          sequence=`expr ${totalLines} - ${M3U8_SIZE}`
+      fi
+      echo "#EXTM3U" > current.m3u8
+      echo "#EXT-X-VERSION:3" >> current.m3u8
+      echo "#EXT-X-TARGETDURATION:${HLS_TIME}" >> current.m3u8
+      echo "#EXT-X-MEDIA-SEQUENCE:${sequence}" >> current.m3u8
+      tail -n ${M3U8_SIZE} ~/process-stream.log | awk '{print "#EXTINF:"$5",\n'${IPFS_GATEWAY}'/ipfs/"$2}' >> current.m3u8
+
+      # Add m3u8 file to IPFS and IPNS publish (uncomment to enable)
+      #m3u8hash=$(ipfs add current.m3u8 | awk '{print $2}')
+      #ipfs name publish --timeout=5s $m3u8hash &
+
+      # Copy files to web server
+      cp current.m3u8 /var/www/html/live.m3u8
+      cp ~/process-stream.log /var/www/html/live.log
     fi
-    echo "#EXTM3U" > current.m3u8
-    echo "#EXT-X-VERSION:3" >> current.m3u8
-    echo "#EXT-X-TARGETDURATION:${HLS_TIME}" >> current.m3u8
-    echo "#EXT-X-MEDIA-SEQUENCE:${sequence}" >> current.m3u8
-    tail -n ${M3U8_SIZE} ~/process-stream.log | awk '{print "#EXTINF:"$5",\n'${IPFS_GATEWAY}'/ipfs/"$2}' >> current.m3u8
-
-    # Add m3u8 file to IPFS
-    m3u8hash=$(ipfs add current.m3u8 | awk '{print $2}')
-
-    # Copy files to web server
-    cp current.m3u8 /var/www/html/live.m3u8
-    cp ~/process-stream.log /var/www/html/live.log
   else
     sleep 5
   fi
